added alert extension to UIView Controller
