--- free-tunes/Utilities/AppAlert.swift
@@ -1,22 +1,0 @@
-//
-//  AppAlert.swift
-//  free-tunes
-//
-//  Created by Tshwarelo Mafaralala on 2022/03/01.
-//
-
-import UIKit
-
-extension UIAlertController {
-
-    func displayErrorAlert(alertTitle: String,
-                           alertMessage: String,
-                           alertActionTitle: String,
-                           alertDelegate: UIViewController) {
-        let alert = UIAlertController(title: alertTitle,
-                                      message: alertMessage,
-                                      preferredStyle: UIAlertController.Style.alert)
-        alert.addAction(UIAlertAction(title: alertActionTitle, style: UIAlertAction.Style.default, handler: nil))
-        alertDelegate.present(alert, animated: true, completion: nil)
-    }
-}
